Test: Refactor tests to only check relevent criteria
The test for creating a valid event was checking that it wasn't having an error, but that doesn't tell us that it was actually successful. We should instead simply check that the status is 201.

The test for viewing all events was similarly checking that there wasn't a specific error, but now checks specifically that the status is 200 and the events aren't empty.