Test: Retrieve the first event created in earlier tests
Test retrieving a specific event by retrieving the first event that was created in the respective test session. The id is generated randomly so there is no way to predict an id. We must use the handler to retrieve it.

Test attempting to retrieve an event with an id that doesn't exist in the system. The response body should be null.